- Added a method for potion usage and a potion bag for mana & health pots (array of integers with a size of 2)

+++ Enemy.java
@@ -27,5 +27,5 @@ public class Enemy extends Character implements Voicelines{
         System.out.println("Hey There!");
     }
 
-    public float getLevel(){ return Level; }
+    public int getLevel(){ return Level; }
 }
